Move binary interpretation into core module
