+++ crates/ddl/src/core/binary.rs
