+++ crates/ddl/src/lib.rs
@@ -1,6 +1,5 @@
 #![warn(rust_2018_idioms)]
 
-pub mod binary;
 pub mod core;
 pub mod diagnostics;
 mod ieee754;
